+++ src/main/java/com/sparta/hhplusconcert/queue/domain/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.sparta.hhplusconcert.domain.queue;
+package com.sparta.hhplusconcert.queue.domain;
 
 public class InvalidTokenException extends RuntimeException{
   public InvalidTokenException(String message) {
