+++ src/main/java/com/sparta/hhplusconcert/queue/domain/entity/QueueTokenEntity.java
@@ -1,4 +1,4 @@
-package com.sparta.hhplusconcert.domain.queue.entity;
+package com.sparta.hhplusconcert.queue.domain.entity;
 
 import com.sparta.hhplusconcert.domain.common.Status;
 import com.sparta.hhplusconcert.domain.common.TimeBaseEntity;
