+++ src/main/java/com/sparta/hhplusconcert/queue/infra/QueueTokenRepository.java
@@ -1,7 +1,7 @@
-package com.sparta.hhplusconcert.infra.queue;
+package com.sparta.hhplusconcert.queue.infra;
 
 import com.sparta.hhplusconcert.domain.common.Status;
-import com.sparta.hhplusconcert.domain.queue.entity.QueueTokenEntity;
+import com.sparta.hhplusconcert.queue.domain.entity.QueueTokenEntity;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
